pyproject.toml
[project]
name = "bolt-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "slack-bolt (>=1.22.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.66.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "bolt_app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
